refactor : Post/U/D 수정,조회 리펙토링

+++ src/main/java/org/team14/newsfeed/service/PostService.java
@@ -1,6 +1,7 @@
 package org.team14.newsfeed.service;
 
 import java.util.List;
+import java.util.ArrayList;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -30,24 +31,27 @@ public class PostService {
     public PostResponseDto createPost(String title, String contents, String username) {
 
         User findUser = userRepository.findUserByUsernameOrElseThrow(username);
+
         Post post = Post.of(title, contents, findUser);
+
         post.setUser(findUser);
 
         postRepository.save(post);
 
-        return new PostResponseDto(post.getId(), post.getTitle(), post.getContents(),
-                post.getUser().getUsername(), post.getUser().getEmail());
-
+        return new PostResponseDto(post.getId(), post.getTitle(), post.getContents(), post.getUser().getUsername(), post.getUser().getEmail());
     }
 
 
-    /*
-        1. 게시글 조회
-        2. 필터(작성자 이메일, 작성자명) 사용
+    /**
+     * 포스트 조회
+     *
+     * @param email 사용자 이메일
+     * @param name  사용자 이름
+     * @return
      */
     public List<PostResponseDto> findPostsByOptionalFilters(String email, String name) {
 
-        List<Post> posts;
+        List<Post> posts = new ArrayList<>();
 
         if (email == null && name == null) {
             // 아무 필터도 존재하지 않음.
@@ -59,6 +63,7 @@ public class PostService {
             // 이메일 필터가 존재할 때. List<Post>
             // 특정 이메일 조회
             posts = postRepository.findAllByUser_Email(email);
+
         } else if (email == null && name != null) {
 
             // 작성자명 필터가 존재할 때.
@@ -69,49 +74,55 @@ public class PostService {
             // 이메일, 작성자명 필터가 존재할 때.
             //예외처리
             posts = postRepository.findAllByUser_EmailAndUser_Username(email, name);
-
         }
 
         if (posts.isEmpty()) {
             throw new CustomException(HttpStatus.NOT_FOUND, "포스트가 없습니다.");
         }
+
         //포스트를 리스트형태로 변환
         return posts.stream()
                 .map(PostResponseDto::toDto)
                 .toList();
     }
 
-    //포스트를 수정하는 메서드
+    /**
+     *포스트를 수정하는 메서드
+     *
+     * @param Id
+     * @param updateRequestDto << 타이틀과 제목을 한번에
+     * @param loggedInEmail 로그인된 이메일
+     * @return
+     */
     @Transactional
-    public PostResponseDto updatePost(Long Id, PostUpdateRequestDto updateRequestDto) {
+    public PostResponseDto updatePost(Long Id, PostUpdateRequestDto updateRequestDto, String loggedInEmail) {
 
         //포스트를 찾기
         Post findPost = postRepository.findByIdOrElseThrow(Id);
-        //타이틀이 널이 아닌 경우 타이틀 수정
-        if (updateRequestDto.getTitle() != null) {
-            findPost.setTitle(updateRequestDto.getTitle());
 
+        // 사용자가 생성한 포스트인지 확인
+        if (!findPost.getUser().getEmail().equals(loggedInEmail)) {
+            throw new CustomException(HttpStatus.BAD_REQUEST,"수정 권한이 없습니다.");
         }
-        //내용이 널이 아닌 경우 널 수정
-        if (updateRequestDto.getContents() != null) {
-            findPost.setContents(updateRequestDto.getContents());
-        }
-        //둘다 널이 아닌 경우 둘다 수정되게 만듦
 
-        return new PostResponseDto(findPost.getId(), findPost.getTitle(), findPost.getContents(),
-                findPost.getUser().getUsername(), findPost.getUser().getEmail());
+        findPost.updateTitleAndContents(updateRequestDto.getTitle(), updateRequestDto.getContents());
+
+        return PostResponseDto.fromEntity(findPost);
     }
 
+    /**
+     * 포스트 삭제
+     * @param Id
+     */
 
-    //삭제 메서드
-    public void delete(Long Id) {
-        //포스트를 찾고
+    public void delete(Long Id, String loggedInEmail) {
 
         Post findPost = postRepository.findByIdOrElseThrow(Id);
 
-        // 이후 삭제
+        if (!findPost.getUser().getEmail().equals(loggedInEmail)) {
+            throw new CustomException(HttpStatus.BAD_REQUEST,"삭제 권한이 없습니다.");
+        }
+
         postRepository.delete(findPost);
     }
-
-
 }
